requirements.txt
aniso8601==3.0.2
certifi==2019.6.16
chardet==3.0.4
decorator==4.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1
django-decouple==2.1
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-su==0.8.0
djangorestframework==3.10.3
djangorestframework-oauth==1.1.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.2.1
graphql-relay==0.4.5
idna==2.7
ipython-genutils==0.2.0
oauthlib==3.1.0
pkg-resources==0.0.0
promise==2.2.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.19.1
requests-oauthlib==1.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
urllib3==1.23
utils==0.9.0